seperate base planning in other operation
