+++ src/constants.js
@@ -19,7 +19,8 @@ module.exports = {
     OPERATION_MAP: 11,
     OPERATION_TOWER: 12,
     OPERATION_HARVESTING: 13,
-    OPERATION_MAX: 13,
+    OPERATION_BASEPLAN: 14,
+    OPERATION_MAX: 14,
 
     BASE_PHASE_BIRTH: 0,
     BASE_PHASE_HARVESTER: 1,
