+++ src/operation.js
@@ -14,6 +14,7 @@ module.exports = class Operation {
         this._id = idIndex++;
         this._runTactics = false;
         this._runStrategy = false;
+        this._runSupport = false;
         /**@type {ChildOp[][]} */
         this._childOps = []
         /**@type {Debug} */
@@ -44,22 +45,25 @@ module.exports = class Operation {
         //last resort cpu overflow prevention.
         if (Game.cpu.bucket < Game.cpu.getUsed() + Game.cpu.limit) throw Error('Out of CPU');
 
-        if (Game.time % SUPPORT_INTERVAL == this._tickOffset) {
+        if (this._runSupport || Game.time % SUPPORT_INTERVAL == this._tickOffset) {
             if(this._debug.verbose) this._debug.logState('support', this)
             try {
                 this._support();
+                if (this._runSupport) this._runSupport = false;
             } catch(err) {this._debug.logError(err)};
         }
         if (this._runStrategy || Game.time % STRATEGY_INTERVAL == this._tickOffset % STRATEGY_INTERVAL) {
             if(this._debug.verbose) this._debug.logState('strategy', this)
             try {
                 this._strategy();
+                if (this._runStrategy) this._runStrategy = false;
             } catch(err) {this._debug.logError(err)};
         }
         if (this._runTactics || Game.time % TACTICS_INTERVAL == this._tickOffset % TACTICS_INTERVAL) {
             if(this._debug.verbose) this._debug.logState('tactics', this)
             try {
                 this._tactics();
+                if (this._runTactics) this._runTactics = false;
             } catch(err) {this._debug.logError(err)};
         }
         if(this._debug.verbose) this._debug.logState('command', this)
@@ -71,8 +75,6 @@ module.exports = class Operation {
                 childOp.run();
             } catch(err) {this._debug.logError(err)}
         }
-        if (this._runTactics) this._runTactics = false;
-        if (this._runStrategy) this._runStrategy = false;
         if(this._debug.verbose) this._debug.logState('end', this)
     }
